Add retries and better display of error messages
